+++ tests/test_wait_signal.py
@@ -111,6 +111,21 @@ def test_signal_triggered(
     stop_watch.check(timeout, delay)
 
 
+@pytest.mark.parametrize("delayed", [True, False])
+def test_zero_timeout(qtbot, timer, delayed, signaller):
+    """
+    With a zero timeout, we don't run a main loop, so only immediate signals are
+    processed.
+    """
+    with qtbot.waitSignal(signaller.signal, raising=False, timeout=0) as blocker:
+        if delayed:
+            timer.single_shot(signaller.signal, 0)
+        else:
+            signaller.signal.emit()
+
+    assert blocker.signal_triggered != delayed
+
+
 @pytest.mark.parametrize(
     "configval, raises", [("false", False), ("true", True), (None, True)]
 )